Refactoring
Converstions of styles
